#193: refactored MyRoles component
